restructured a bit into extension functions
