+++ src/archerly.api/Program.cs
@@ -2,17 +2,13 @@ namespace archerly.api;
 
 using Serilog;
 using Serilog.Sinks.Loki;
-using Prometheus;
+using archerly.api.extensions;
 
 public static class Program
 {
     public static void Main(string[] args)
     {
-
         var builder = WebApplication.CreateBuilder(args);
-        // replace with actual port for metrics
-        // using var server = new KestrelMetricServer(port: 1234);
-        // server.Start();
 
         // Add services to the container.
         // Learn more about configuring OpenAPI at https://aka.ms/aspnet/openapi
@@ -24,10 +20,10 @@ public static class Program
         //var credentials = new NoAuthCredentials("http://localhost:3100"); 
 
         Log.Logger = new LoggerConfiguration()
-        .MinimumLevel.Information()
-        .Enrich.FromLogContext()
-        .WriteTo.LokiHttp(credentials)
-        .CreateLogger();
+                    .MinimumLevel.Information()
+                    .Enrich.FromLogContext()
+                    .WriteTo.LokiHttp(credentials)
+                    .CreateLogger();
 
         var app = builder.Build();
 
@@ -38,25 +34,12 @@ public static class Program
         }
 
         app.UseHttpsRedirection();
+
         // set up metrics instrumentation
-        app.UseMetricServer();
-        app.UseHttpMetrics();
+        app.UseMetrics();
 
-        var summaries = new[] { "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching" };
-
-        app.MapGet("/weatherforecast", () =>
-        {
-            var forecast = Enumerable.Range(1, 5).Select(index =>
-                new WeatherForecast
-                (
-                    DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-                    Random.Shared.Next(-20, 55),
-                    summaries[Random.Shared.Next(summaries.Length)]
-                ))
-                .ToArray();
-            return forecast;
-        })
-        .WithName("GetWeatherForecast");
+        // define all api routes
+        app.UseRoutes();
 
         app.Run();
     }
